+++ src/test/php/webservices/rest/unittest/srv/ResponseTest.class.php
@@ -1,101 +1,65 @@
 <?php namespace webservices\rest\unittest\srv;
 
-
-
-use unittest\TestCase;
 use webservices\rest\srv\Response;
-
+use webservices\rest\Payload;
+use scriptlet\Cookie;
+use scriptlet\HttpScriptletResponse;
+use peer\URL;
 
 /**
  * Test response class
  *
  * @see  xp://webservices.rest.srv.Response
  */
-class ResponseTest extends TestCase {
+class ResponseTest extends \unittest\TestCase {
 
-  /**
-   * Test constructor
-   * 
-   */
   #[@test]
   public function create() {
-    $this->assertEquals(null, create(new Response())->status);
+    $this->assertEquals(null, (new Response())->status);
   }
 
-  /**
-   * Test constructor
-   * 
-   */
   #[@test]
   public function create_with_status() {
-    $this->assertEquals(200, create(new Response(200))->status);
+    $this->assertEquals(200, (new Response(200))->status);
   }
 
-  /**
-   * Test payload is initially null
-   * 
-   */
   #[@test]
   public function payload_initially_null() {
-    $this->assertNull(create(new Response())->payload);
+    $this->assertNull((new Response())->payload);
   }
 
-  /**
-   * Test headers are initially empty
-   * 
-   */
   #[@test]
   public function headers_initially_empty() {
-    $this->assertEquals(array(), create(new Response())->headers);
+    $this->assertEquals([], (new Response())->headers);
   }
 
-  /**
-   * Test ok() method
-   * 
-   */
   #[@test]
   public function ok() {
     $r= Response::ok();
     $this->assertEquals(200, $r->status);
   }
 
-  /**
-   * Test created() method
-   * 
-   */
   #[@test]
   public function created() {
     $r= Response::created();
     $this->assertEquals(201, $r->status);
-    $this->assertEquals(array(), $r->headers);
+    $this->assertEquals([], $r->headers);
   }
 
-  /**
-   * Test created() method
-   * 
-   */
   #[@test]
   public function created_with_location() {
     $location= 'http://example.com/resource/4711';
     $r= Response::created($location);
     $this->assertEquals(201, $r->status);
-    $this->assertEquals(array('Location' => $location), $r->headers);
+    $this->assertEquals(['Location' => $location], $r->headers);
   }
 
-  /**
-   * Test noContent() method
-   * 
-   */
   #[@test]
   public function no_content() {
     $r= Response::noContent();
     $this->assertEquals(204, $r->status);
   }
 
-  /**
-   * Test see() method
-   * 
-   */
   #[@test]
   public function see() {
     $location= 'http://example.com/resource/4711';
@@ -103,121 +67,97 @@ class ResponseTest extends TestCase {
     $this->assertEquals(302, $r->status);
   }
 
-  /**
-   * Test notModified() method
-   * 
-   */
   #[@test]
   public function not_modified() {
     $r= Response::notModified();
     $this->assertEquals(304, $r->status);
   }
 
-  /**
-   * Test notFound() method
-   * 
-   */
   #[@test]
   public function not_found() {
     $r= Response::notFound();
     $this->assertEquals(404, $r->status);
   }
 
-  /**
-   * Test notAcceptable() method
-   * 
-   */
   #[@test]
   public function not_acceptable() {
     $r= Response::notAcceptable();
     $this->assertEquals(406, $r->status);
   }
 
-  /**
-   * Test error() method
-   * 
-   */
   #[@test]
   public function error() {
     $r= Response::error();
     $this->assertEquals(500, $r->status);
   }
 
-  /**
-   * Test error() method
-   * 
-   */
   #[@test]
   public function error_503() {
     $r= Response::error(503);
     $this->assertEquals(503, $r->status);
   }
 
-  /**
-   * Test status() method
-   * 
-   */
   #[@test]
   public function status_402() {
     $r= Response::status(402);
     $this->assertEquals(402, $r->status);
   }
 
-  /**
-   * Test withHeader() method
-   * 
-   */
+  #[@test]
+  public function not_found_with_message() {
+    $r= Response::notFound('No file named four-oh-four');
+    $this->assertEquals(new Payload('No file named four-oh-four'), $r->payload);
+  }
+
+  #[@test]
+  public function not_acceptable_with_message() {
+    $r= Response::notAcceptable('Cannot upload files named four-our-six');
+    $this->assertEquals(new Payload('Cannot upload files named four-our-six'), $r->payload);
+  }
+
+  #[@test]
+  public function error_with_message() {
+    $r= Response::error(503, 'Come back later');
+    $this->assertEquals(new Payload('Come back later'), $r->payload);
+  }
+
+  #[@test]
+  public function status_with_message() {
+    $r= Response::status(203, 'Eventually consistent');
+    $this->assertEquals(new Payload('Eventually consistent'), $r->payload);
+  }
+
   #[@test]
   public function with_extra_header() {
-    $r= create(new Response())->withHeader('X-Exception', 'SQL');
-    $this->assertEquals(array('X-Exception' => 'SQL'), $r->headers);
+    $r= (new Response())->withHeader('X-Exception', 'SQL');
+    $this->assertEquals(['X-Exception' => 'SQL'], $r->headers);
   }
 
-  /**
-   * Test withPayload() method
-   * 
-   */
   #[@test]
   public function with_payload() {
-    $data= array('name' => 'example');
-    $r= create(new Response())->withPayload($data);
-    $this->assertEquals(new \webservices\rest\Payload($data), $r->payload);
+    $data= ['name' => 'example'];
+    $r= (new Response())->withPayload($data);
+    $this->assertEquals(new Payload($data), $r->payload);
   }
 
-  /**
-   * Test withPayload() method
-   * 
-   */
   #[@test]
   public function with_payload_instance() {
-    $data= array('name' => 'example');
-    $r= create(new Response())->withPayload(new \webservices\rest\Payload($data));
-    $this->assertEquals(new \webservices\rest\Payload($data), $r->payload);
+    $data= ['name' => 'example'];
+    $r= (new Response())->withPayload(new Payload($data));
+    $this->assertEquals(new Payload($data), $r->payload);
   }
 
-  /**
-   * Test equals() method
-   * 
-   */
   #[@test]
   public function equals_identical() {
     $r= Response::status(200);
     $this->assertEquals($r, $r);
   }
 
-  /**
-   * Test equals() method
-   * 
-   */
   #[@test]
   public function equals_same() {
     $this->assertEquals(Response::status(200), Response::status(200));
   }
 
-  /**
-   * Test equals() method
-   * 
-   */
   #[@test]
   public function equals_with_headers() {
     $this->assertEquals(
@@ -226,19 +166,11 @@ class ResponseTest extends TestCase {
     );
   }
 
-  /**
-   * Test equals() method
-   * 
-   */
   #[@test]
   public function equals_different_status() {
     $this->assertNotEquals(Response::status(200), Response::status(201));
   }
 
-  /**
-   * Test equals() method
-   * 
-   */
   #[@test]
   public function equals_different_header_values() {
     $this->assertNotEquals(
@@ -247,10 +179,6 @@ class ResponseTest extends TestCase {
     );
   }
 
-  /**
-   * Test equals() method
-   * 
-   */
   #[@test]
   public function equals_different_header_keys() {
     $this->assertNotEquals(
@@ -259,10 +187,6 @@ class ResponseTest extends TestCase {
     );
   }
 
-  /**
-   * Test equals() method
-   * 
-   */
   #[@test]
   public function equals_different_header_sizes() {
     $this->assertNotEquals(
@@ -271,10 +195,6 @@ class ResponseTest extends TestCase {
     );
   }
 
-  /**
-   * Test equals() method
-   * 
-   */
   #[@test]
   public function equals_same_primitive_payloads() {
     $this->assertEquals(
@@ -283,10 +203,6 @@ class ResponseTest extends TestCase {
     );
   }
 
-  /**
-   * Test equals() method
-   * 
-   */
   #[@test]
   public function equals_different_primitive_payloads() {
     $this->assertNotEquals(
@@ -295,10 +211,6 @@ class ResponseTest extends TestCase {
     );
   }
 
-  /**
-   * Test equals() method
-   * 
-   */
   #[@test]
   public function equals_same_array_payloads() {
     $this->assertEquals(
@@ -307,10 +219,6 @@ class ResponseTest extends TestCase {
     );
   }
 
-  /**
-   * Test equals() method
-   * 
-   */
   #[@test]
   public function equals_different_array_payloads() {
     $this->assertNotEquals(
@@ -319,10 +227,6 @@ class ResponseTest extends TestCase {
     );
   }
 
-  /**
-   * Test equals() method
-   * 
-   */
   #[@test]
   public function equals_identical_object_payloads() {
     $this->assertEquals(
@@ -331,13 +235,9 @@ class ResponseTest extends TestCase {
     );
   }
 
-  /**
-   * Test equals() method
-   * 
-   */
   #[@test]
   public function equals_same_object_payloads() {
-    $class= \lang\ClassLoader::defineClass('ResponseTest_SameObjectFixture', 'lang.Object', array(), '{
+    $class= \lang\ClassLoader::defineClass('ResponseTest_SameObjectFixture', 'lang.Object', [], '{
       public function equals($cmp) { return TRUE; }
     }');
     $this->assertEquals(
@@ -346,10 +246,6 @@ class ResponseTest extends TestCase {
     );
   }
 
-  /**
-   * Test equals() method
-   * 
-   */
   #[@test]
   public function equals_different_object_payloads() {
     $this->assertNotEquals(
@@ -358,10 +254,6 @@ class ResponseTest extends TestCase {
     );
   }
 
-  /**
-   * Test equals() method
-   * 
-   */
   #[@test]
   public function equals_object_and_null_payloads() {
     $this->assertNotEquals(
@@ -370,10 +262,6 @@ class ResponseTest extends TestCase {
     );
   }
 
-  /**
-   * Test equals() method
-   * 
-   */
   #[@test]
   public function equals_null() {
     $this->assertEquals(
@@ -382,56 +270,44 @@ class ResponseTest extends TestCase {
     );
   }
 
-  /**
-   * Test cookies
-   * 
-   */
   #[@test]
   public function without_cookies() {
     $this->assertEquals(
-      array(),
+      [],
       Response::status(200)->cookies 
     );
   }
 
-  /**
-   * Test cookies
-   * 
-   */
   #[@test]
   public function with_one_cookie() {
-    $user= new \scriptlet\Cookie('user', 'Test');
+    $user= new Cookie('user', 'Test');
     $this->assertEquals(
-      array($user),
+      [$user],
       Response::status(200)->withCookie($user)->cookies 
     );
   }
 
-  /**
-   * Test cookies
-   * 
-   */
   #[@test]
   public function with_two_cookies() {
-    $user= new \scriptlet\Cookie('user', 'Test');
-    $lang= new \scriptlet\Cookie('language', 'de');
+    $user= new Cookie('user', 'Test');
+    $lang= new Cookie('language', 'de');
     $this->assertEquals(
-      array($user, $lang),
+      [$user, $lang],
       Response::status(200)->withCookie($user)->withCookie($lang)->cookies 
     );
   }
 
   #[@test]
   public function writeTo_fully_qualifies_path_in_location_header() {
-    $res= new \scriptlet\HttpScriptletResponse();
-    Response::see('/foo')->writeTo($res, new \peer\URL('http://example.com/'), null);
+    $res= new HttpScriptletResponse();
+    Response::see('/foo')->writeTo($res, new URL('http://example.com/'), null);
     $this->assertEquals('Location: http://example.com/foo', $res->headers[0]);
   }
 
   #[@test]
   public function writeTo_does_not_fully_qualify_url_in_location_header() {
-    $res= new \scriptlet\HttpScriptletResponse();
-    Response::see('http://localhost/')->writeTo($res, new \peer\URL('http://example.com/'), null);
+    $res= new HttpScriptletResponse();
+    Response::see('http://localhost/')->writeTo($res, new URL('http://example.com/'), null);
     $this->assertEquals('Location: http://localhost/', $res->headers[0]);
   }
 }